Fix swapped test file names
