+++ src/pages/FeaturePreviews.test.ts
@@ -1,41 +1,76 @@
-import { act, render, screen } from '@testing-library/react';
-import userEvent from '@testing-library/user-event';
+import { fireEvent, getByText, render } from '@testing-library/react';
 import { h } from 'react-hyperscript-helpers';
-import { PauseButton } from 'src/pages/Environments';
-import { cloudProviders } from 'src/pages/workspaces/workspace/analysis/utils/runtime-utils';
+import { isFeaturePreviewEnabled, toggleFeaturePreview, useAvailableFeaturePreviews } from 'src/libs/feature-previews';
+import { FeaturePreviews } from 'src/pages/FeaturePreviews';
 
-describe('Environments', () => {
-  it.each([
-    { cloudContext: cloudProviders.gcp.label, label: 'Cromwell' },
-    { cloudContext: cloudProviders.azure.label, label: 'CromwellOnAzure' },
-  ])('should enable pause for azure and google', async ({ cloudContext, label }) => {
-    // Arrange
-    const pauseComputeAndRefresh = jest.fn();
-    const testCompute = {
-      appName: 'test-app',
-      cloudContext,
-      kubernetesRuntimeConfig: { numNodes: 1, machineType: 'n1-highmem-8', autoscalingEnabled: false },
-      errors: [],
-      status: 'RUNNING',
-      proxyUrls: {},
-      customEnvironmentVariables: {},
-      auditInfo: {
-        creator: 'Jake',
-        createdDate: '2022-01-24T15:27:28.740880Z',
-        dateAccessed: '2022-02-24T15:27:28.740880Z',
-      },
-      appType: label,
-      labels: { tool: '' },
-    };
+jest.mock('src/libs/ajax');
+jest.mock('src/libs/feature-previews');
 
-    await act(async () => { render(h(PauseButton, { computeType: 'app', cloudEnvironment: testCompute, currentUser: 'Jake', pauseComputeAndRefresh })) }) //eslint-disable-line
-    // Act
-    const pauseButton = screen.getByText('Pause');
-    // Assert
-    expect(pauseButton).toBeEnabled();
-    // Act
-    await userEvent.click(pauseButton);
-    // Assert
-    expect(pauseComputeAndRefresh).toHaveBeenCalled();
+describe('FeaturePreviews', () => {
+  beforeEach(() => {
+    useAvailableFeaturePreviews.mockReturnValue({
+      featurePreviews: [
+        {
+          id: 'feature1',
+          title: 'Feature #1',
+          description: 'A new feature',
+          documentationUrl: 'https://example.com/feature-1-docs',
+        },
+        {
+          id: 'feature2',
+          title: 'Feature #2',
+          description: 'Another new feature',
+          feedbackUrl: 'mailto:feature2-feedback@example.com',
+        },
+      ],
+    });
+
+    isFeaturePreviewEnabled.mockReturnValue(false);
+  });
+
+  it('should render available feature previews', () => {
+    const { getAllByRole } = render(h(FeaturePreviews));
+    const cells = getAllByRole('cell');
+
+    expect(getByText(cells[1], 'Feature #1')).toBeTruthy();
+    expect(getByText(cells[1], 'A new feature')).toBeTruthy();
+
+    expect(getByText(cells[3], 'Feature #2')).toBeTruthy();
+    expect(getByText(cells[3], 'Another new feature')).toBeTruthy();
+  });
+
+  it('should render whether features are enabled', () => {
+    isFeaturePreviewEnabled.mockImplementation((id) => id === 'feature1');
+
+    const { getAllByRole } = render(h(FeaturePreviews));
+    const checkboxes = getAllByRole('checkbox');
+
+    expect(checkboxes[0].getAttribute('aria-checked')).toBe('true');
+    expect(checkboxes[1].getAttribute('aria-checked')).toBe('false');
+  });
+
+  it('checking a checkbox should toggle feature previews', () => {
+    const { getAllByRole } = render(h(FeaturePreviews));
+    const checkboxes = getAllByRole('checkbox');
+
+    fireEvent.click(checkboxes[0]);
+    expect(toggleFeaturePreview).toHaveBeenCalledWith('feature1', true);
+
+    fireEvent.click(checkboxes[0]);
+    expect(toggleFeaturePreview).toHaveBeenCalledWith('feature1', false);
+  });
+
+  it('should render documentation link if provided', () => {
+    const { getAllByText } = render(h(FeaturePreviews));
+    const docLinks = getAllByText('Documentation');
+    expect(docLinks.length).toBe(1);
+    expect(docLinks[0].getAttribute('href')).toBe('https://example.com/feature-1-docs');
+  });
+
+  it('should render feedback link if provided', () => {
+    const { getAllByText } = render(h(FeaturePreviews));
+    const feedbackLinks = getAllByText('Submit feedback');
+    expect(feedbackLinks.length).toBe(1);
+    expect(feedbackLinks[0].getAttribute('href')).toBe('mailto:feature2-feedback@example.com');
   });
 });
